Replaced old model with new.
