+++ glotaran/model/interval_property.py
@@ -1,23 +1,12 @@
-"""Helper functions."""
 from __future__ import annotations
 
-from typing import List
-from typing import Tuple
-
-from glotaran.model.item import model_item
+from glotaran.model.item import Item
+from glotaran.model.item import item
 
 
-@model_item(
-    properties={
-        "interval": {"type": List[Tuple[float, float]], "default": None, "allow_none": True},
-    },
-    has_label=False,
-)
-class IntervalProperty:
-    """Applies a relation between clps as
-
-    :math:`source = parameter * target`.
-    """
+@item
+class IntervalProperty(Item):
+    interval: tuple[float, float] | list[tuple[float, float]] | None = None
 
     def has_interval(self) -> bool:
         return self.interval is not None
@@ -35,11 +24,14 @@ class IntervalProperty:
         applies : bool
 
         """
-        if self.interval is None:
+        if self.interval is None or value is None:
             return True
 
-        def applies(interval):
-            return interval[0] <= value <= interval[1]
+        def applies(interval: tuple[float, float]):
+            lower, upper = interval[0], interval[1]
+            if lower > upper:
+                lower, upper = upper, lower
+            return lower <= value <= upper
 
         if isinstance(self.interval, tuple):
             return applies(self.interval)
