+++ glotaran/model/item.py
@@ -1,393 +1,312 @@
-"""The model item decorator."""
 from __future__ import annotations
 
-import copy
+from inspect import getmro
 from textwrap import indent
+from types import NoneType
+from types import UnionType
 from typing import TYPE_CHECKING
 from typing import Callable
-from typing import List
+from typing import ClassVar
+from typing import Generator
+from typing import Iterator
 from typing import Type
+from typing import TypeAlias
+from typing import TypeVar
+from typing import Union
+from typing import get_args
+from typing import get_origin
 
-from glotaran.model.property import ModelProperty
-from glotaran.model.util import wrap_func_as_method
+from attrs import NOTHING
+from attrs import Attribute
+from attrs import define
+from attrs import field
+from attrs import fields
+from attrs import resolve_types
+
+from glotaran.parameter import Parameter
+from glotaran.parameter import ParameterGroup
 from glotaran.utils.ipython import MarkdownStr
 
 if TYPE_CHECKING:
-    from typing import Any
-
     from glotaran.model.model import Model
-    from glotaran.parameter import ParameterGroup
-
-    Validator = Callable[
-        [Type[object], Type[Model]],
-        List[str],
-    ]
-
-    ValidatorParameter = Callable[
-        [Type[object], Type[Model], Type[ParameterGroup]],
-        List[str],
-    ]
 
 
-def model_item(
-    properties: None | dict[str, dict[str, Any]] = None,
-    has_type: bool = False,
-    has_label: bool = True,
-) -> Callable:
-    """The `@model_item` decorator adds the given properties to the class. Further it adds
-    classmethods for deserialization, validation and printing.
+META_ALIAS = "__glotaran_alias__"
+META_VALIDATOR = "__glotaran_validator__"
 
-    By default, a `label` property is added.
 
-    The `properties` dictionary contains the name of the properties as keys. The values must be
-    either a `type` or dictionary with the following values:
+class ItemIssue:
+    def to_string(self) -> str:
+        raise NotImplementedError
 
-    * type: a `type` (required)
-    * doc: a string for documentation (optional)
-    * default: a default value (optional)
-    * allow_none: if `True`, the property can be set to None (optional)
+    def __rep__(self) -> str:
+        return self.to_string()
 
-    Classes with the `model_item` decorator intended to be used in glotaran models.
 
-    Parameters
-    ----------
-    properties :
-        A dictionary of property names and options.
-    has_type :
-        If true, a type property will added. Used for model attributes, which
-        can have more then one type.
-    has_label :
-        If false no label property will be added.
-    """
+class ModelItemIssue(ItemIssue):
+    def __init__(self, item_name: str, label: str):
+        self._item_name = item_name
+        self._label = label
 
-    if properties is None:
-        properties = {}
+    def to_string(self) -> str:
+        return f"Missing model item '{self.item_name}' with label '{self.label}'."
 
-    def decorator(cls):
 
-        setattr(cls, "_glotaran_has_label", has_label)
-        setattr(cls, "_glotaran_model_item", True)
+class ParameterIssue(ItemIssue):
+    def __init__(self, label: str):
+        self._label = label
 
-        # store for later sanity checking
-        if not hasattr(cls, "_glotaran_properties"):
-            setattr(cls, "_glotaran_properties", [])
-            if has_label:
-                doc = f"The label of {cls.__name__} item."
-                prop = ModelProperty(cls, "label", str, doc, None, False)
-                setattr(cls, "label", prop)
-                getattr(cls, "_glotaran_properties").append("label")
-            if has_type:
-                doc = f"The type string of {cls.__name__}."
-                prop = ModelProperty(cls, "type", str, doc, None, True)
-                setattr(cls, "type", prop)
-                getattr(cls, "_glotaran_properties").append("type")
+    def to_string(self) -> str:
+        return f"Missing parameter with label '{self.label}'."
+
+
+class Item:
+    pass
+
+
+def item_to_markdown(
+    item: Item, parameters: ParameterGroup = None, initial_parameters: ParameterGroup = None
+) -> MarkdownStr:
+    md = "\n"
+    for attr in fields(item.__class__):
+        name = attr.name
+        value = getattr(item, name)
+        if value is None:
+            continue
+
+        structure, item_type = strip_type_and_structure_from_attribute(attr)
+        if item_type is Parameter and parameters is not None:
+            if structure is dict:
+                value = {
+                    k: parameters.get(v).markdown(parameters, initial_parameters)
+                    for k, v in value.items()
+                }
+            elif structure is list:
+                value = [parameters.get(v).markdown(parameters, initial_parameters) for v in value]
+            else:
+                value = parameters.get(value).markdown(parameters, initial_parameters)
+
+        property_md = indent(f"* *{name.replace('_', ' ').title()}*: {value}\n", "  ")
+
+        md += property_md
+
+    return MarkdownStr(md)
+
+
+def iterate_attributes_of_type(item: Item, attr_type: type) -> Generator[Attribute, None, None]:
+    for attr in fields(item):
+        _, item_type = strip_type_and_structure_from_attribute(attr)
+        if isinstance(item_type, Type) and issubclass(item_type, attr_type):
+            yield attr
+
+
+def model_attributes(item: Item, with_alias: bool = True) -> Generator[Attribute, None, None]:
+    for attr in iterate_attributes_of_type(item, ModelItem):
+        if with_alias or META_ALIAS not in attr.metadata:
+            yield attr
+
+
+def parameter_attributes(item: Item) -> Generator[Attribute, None, None]:
+    return iterate_attributes_of_type(item, Parameter)
+
+
+def iterate_names_and_labels(
+    item: Item, attributes: Generator[Attribute, None, None]
+) -> Generator[tuple(str, str), None, None]:
+    for attr in attributes:
+        structure, _ = strip_type_and_structure_from_attribute(attr)
+        value = getattr(item, attr.name)
+        name = attr.metadata.get(META_ALIAS, attr.name)
+
+        if not value:
+            continue
+
+        if structure is dict:
+            for v in value.values():
+                yield name, v
+
+        elif structure is list:
+            for v in value:
+                yield name, v
 
         else:
-            setattr(
-                cls,
-                "_glotaran_properties",
-                list(getattr(cls, "_glotaran_properties")),
-            )
-
-        for name, options in properties.items():
-            if not isinstance(options, dict):
-                options = {"type": options}
-            prop = ModelProperty(
-                cls,
-                name,
-                options.get("type"),
-                options.get("doc", f"{name}"),
-                options.get("default", None),
-                options.get("allow_none", False),
-            )
-            setattr(cls, name, prop)
-            if name not in getattr(cls, "_glotaran_properties"):
-                getattr(cls, "_glotaran_properties").append(name)
-
-        validators = _get_validators(cls)
-        setattr(cls, "_glotaran_validators", validators)
-
-        init = _init_factory(cls)
-        setattr(cls, "__init__", init)
-
-        from_dict = _from_dict_factory(cls)
-        setattr(cls, "from_dict", from_dict)
-
-        validate = _validation_factory(cls)
-        setattr(cls, "validate", validate)
-
-        as_dict = _as_dict_factory(cls)
-        setattr(cls, "as_dict", as_dict)
-
-        get_state = _get_state_factory(cls)
-        setattr(cls, "__getstate__", get_state)
-
-        set_state = _set_state_factory(cls)
-        setattr(cls, "__setstate__", set_state)
-
-        fill = _fill_factory(cls)
-        setattr(cls, "fill", fill)
-
-        markdown = _markdown_factory(cls)
-        setattr(cls, "markdown", markdown)
-
-        get_parameter_labels = _get_parameter_labels_factory(cls)
-        setattr(cls, "get_parameter_labels", get_parameter_labels)
-
-        return cls
-
-    return decorator
+            yield name, value
 
 
-def model_item_typed(
+def iterate_model_item_names_and_labels(item: Item) -> Generator[tuple(str, str), None, None]:
+    return iterate_names_and_labels(item, model_attributes(item.__class__))
+
+
+def iterate_parameter_names_and_labels(item: Item) -> Generator[tuple(str, str), None, None]:
+    return iterate_names_and_labels(item, parameter_attributes(item.__class__))
+
+
+def fill_item_attributes(
+    item: item, iterator: Iterator[Attribute], fill_function: Callable
+) -> list[ItemIssue]:
+    for attr in iterator:
+        value = getattr(item, attr.name)
+
+        structure, _ = strip_type_and_structure_from_attribute(attr)
+        name = attr.metadata.get(META_ALIAS, attr.name)
+        if structure is dict:
+            value = {
+                k: fill_function(name, v) if isinstance(v, str) else v for k, v in value.items()
+            }
+        elif structure is list:
+            value = [fill_function(name, v) if isinstance(v, str) else v for v in value]
+        else:
+            value = fill_function(name, value) if isinstance(value, str) else value
+
+        setattr(item, attr.name, value)
+
+
+def fill_item(item: item, model: Model, parameters: ParameterGroup) -> Item:
+    fill_item_model_attributes(item, model, parameters)
+    fill_item_parameter_attributes(item, parameters)
+    return item
+
+
+def fill_item_model_attributes(item: item, model: Model, parameters: ParameterGroup):
+    fill_item_attributes(
+        item,
+        model_attributes(item.__class__),
+        lambda name, label: fill_item(getattr(model, name)[label], model, parameters),
+    )
+
+
+def fill_item_parameter_attributes(item: item, parameters: ParameterGroup) -> list[ItemIssue]:
+    fill_item_attributes(
+        item, parameter_attributes(item.__class__), lambda name, label: parameters.get(label)
+    )
+
+
+def get_item_model_issues(item: item, model: Model) -> list[ItemIssue]:
+    return [
+        ModelItemIssue(name, label)
+        for name, label in iterate_model_item_names_and_labels(item)
+        if label not in getattr(model, name)
+    ]
+
+
+def get_item_parameter_issues(item: item, parameters: ParameterGroup) -> list[ItemIssue]:
+    return [
+        ParameterIssue(label)
+        for name, label in iterate_parameter_names_and_labels(item)
+        if not parameters.has(label)
+    ]
+
+
+def get_item_validator_issues(
+    item: item, model: Model, parameters: ParameterGroup | None = None
+) -> list[ItemIssue]:
+    issues = []
+    for name, validator in [
+        (attr.name, attr.metadata[META_VALIDATOR])
+        for attr in fields(item.__class__)
+        if META_VALIDATOR in attr.metadata
+    ]:
+        issues += validator(getattr(item, name), model, parameters)
+
+    return issues
+
+
+def get_item_issues(
+    *, item: Item, model: Model, parameters: ParameterGroup | None = None
+) -> list[ItemIssue]:
+    issues = get_item_model_issues(item, model)
+    issues += get_item_validator_issues(item, model, parameters)
+    if parameters is not None:
+        issues += get_item_parameter_issues(item, parameters)
+    return issues
+
+
+@define(kw_only=True, slots=False)
+class ModelItem(Item):
+    label: str
+
+
+T = TypeVar("ModelItemT", bound="ModelItem")
+
+ParameterType: TypeAlias = Parameter | str
+ModelItemType: TypeAlias = T | str
+
+
+@define(kw_only=True, slots=False)
+class TypedItem(Item):
+    type: str
+    __model_item_types__: ClassVar[dict[str, type]]
+
+    @classmethod
+    def _register_item_class(cls):
+        item_type = cls.get_item_type()
+        if item_type is not NOTHING:
+            cls.__model_item_types__[item_type] = cls
+
+    @classmethod
+    def get_item_type(cls) -> str:
+        return fields(cls).type.default
+
+    @classmethod
+    def get_item_types(cls) -> str:
+        return cls.__model_item_types__.keys()
+
+    @classmethod
+    def get_item_type_class(cls, item_type: str) -> type:
+        return cls.__model_item_types__[item_type]
+
+
+@define(kw_only=True, slots=False)
+class ModelItemTyped(TypedItem, ModelItem):
+    pass
+
+
+def strip_type_and_structure_from_attribute(attr: Attribute) -> tuple[None | list | dict, type]:
+    definition = attr.type
+    definition = strip_option_type(definition)
+    structure, definition = strip_struture_type(definition)
+    definition = strip_option_type(definition, strip_type=str)
+    return structure, definition
+
+
+def strip_option_type(definition: type, strip_type: type = NoneType) -> type:
+    if get_origin(definition) in [Union, UnionType] and strip_type in get_args(definition):
+        definition = get_args(definition)[0]
+    return definition
+
+
+def strip_struture_type(definition: type) -> tuple[None | list | dict, type]:
+    structure = get_origin(definition)
+    if structure is list:
+        definition = get_args(definition)[0]
+    elif structure is dict:
+        definition = get_args(definition)[1]
+    else:
+        structure = None
+
+    return structure, definition
+
+
+def item(cls):
+    parent = getmro(cls)[1]
+    cls = define(kw_only=True, slots=False)(cls)
+    if parent in (TypedItem, ModelItemTyped):
+        cls.__model_item_types__ = {}
+    elif issubclass(cls, TypedItem):
+        cls._register_item_class()
+    resolve_types(cls)
+    return cls
+
+
+def attribute(
     *,
-    types: dict[str, Any],
-    has_label: bool = True,
-    default_type: str = None,
-):
-    """The model_item_typed decorator adds attributes to the class to enable
-    the glotaran model parser to infer the correct class for an item when there
-    are multiple variants.
-
-    Parameters
-    ----------
-    types :
-        A dictionary of types and options.
-    has_label:
-        If `False` no label property will be added.
-    """
-
-    def decorator(cls):
-
-        setattr(cls, "_glotaran_model_item", True)
-        setattr(cls, "_glotaran_model_item_typed", True)
-        setattr(cls, "_glotaran_model_item_types", types)
-        setattr(cls, "_glotaran_model_item_default_type", default_type)
-
-        get_default_type = _get_default_type_factory(cls)
-        setattr(cls, "get_default_type", get_default_type)
-
-        add_type = _add_type_factory(cls)
-        setattr(cls, "add_type", add_type)
-
-        setattr(cls, "_glotaran_has_label", has_label)
-
-        return cls
-
-    return decorator
-
-
-def model_item_validator(need_parameter: bool):
-    """The model_item_validator marks a method of a model item as validation function"""
-
-    def decorator(method: Validator | ValidatorParameter):
-        setattr(method, "_glotaran_validator", need_parameter)
-        return method
-
-    return decorator
-
-
-def _get_validators(cls):
-    return {
-        method: getattr(getattr(cls, method), "_glotaran_validator")
-        for method in dir(cls)
-        if hasattr(getattr(cls, method), "_glotaran_validator")
-    }
-
-
-def _get_default_type_factory(cls):
-    @classmethod
-    @wrap_func_as_method(cls)
-    def get_default_type(cls) -> str:
-        return getattr(cls, "_glotaran_model_item_default_type")
-
-    return get_default_type
-
-
-def _add_type_factory(cls):
-    @classmethod
-    @wrap_func_as_method(cls)
-    def add_type(cls, type_name: str, attribute_type: type):
-        getattr(cls, "_glotaran_model_item_types")[type_name] = attribute_type
-
-    return add_type
-
-
-def _init_factory(cls):
-    @classmethod
-    @wrap_func_as_method(cls)
-    def __init__(self):
-        for attr in self._glotaran_properties:
-            setattr(self, f"_{attr}", None)
-
-    return __init__
-
-
-def _from_dict_factory(cls):
-    @classmethod
-    @wrap_func_as_method(cls)
-    def from_dict(ncls, values: dict) -> cls:
-        f"""Creates an instance of {cls.__name__} from a dictionary of values.
-
-        Intended only for internal use.
-
-        Parameters
-        ----------
-        values :
-            A list of values.
-        """
-        item = ncls()
-
-        for name in ncls._glotaran_properties:
-            if name in values:
-                value = values[name]
-                prop = getattr(item.__class__, name)
-                if prop.glotaran_property_type == float:
-                    value = float(value)
-                elif prop.glotaran_property_type == int:
-                    value = int(value)
-                setattr(item, name, value)
-
-            elif not getattr(ncls, name).glotaran_allow_none and getattr(item, name) is None:
-                raise ValueError(f"Missing Property '{name}' For Item '{ncls.__name__}'")
-        return item
-
-    return from_dict
-
-
-def _validation_factory(cls):
-    @wrap_func_as_method(cls)
-    def validate(self, model: Model, parameters: ParameterGroup | None = None) -> list[str]:
-        f"""Creates a list of parameters needed by this instance of {cls.__name__} not present in a
-        set of parameters.
-
-        Parameters
-        ----------
-        model :
-            The model to validate.
-        parameter :
-            The parameter to validate.
-        missing :
-            A list the missing will be appended to.
-        """
-        problems = []
-        for name in self._glotaran_properties:
-            prop = getattr(self.__class__, name)
-            value = getattr(self, name)
-            problems += prop.glotaran_validate(value, model, parameters)
-        for validator, need_parameter in self._glotaran_validators.items():
-            if need_parameter:
-                if parameters is not None:
-                    problems += getattr(self, validator)(model, parameters)
-            else:
-                problems += getattr(self, validator)(model)
-
-        return problems
-
-    return validate
-
-
-def _as_dict_factory(cls):
-    @wrap_func_as_method(cls)
-    def as_dict(self) -> dict:
-        return {
-            name: getattr(self.__class__, name).glotaran_replace_parameter_with_labels(
-                getattr(self, name)
-            )
-            for name in self._glotaran_properties
-            if name != "label" and getattr(self, name) is not None
-        }
-
-    return as_dict
-
-
-def _fill_factory(cls):
-    @wrap_func_as_method(cls)
-    def fill(self, model: Model, parameters: ParameterGroup) -> cls:
-        f"""Returns a copy of the {cls.__name__} instance with all members which are Parameters are
-        replaced by the value of the corresponding parameter in the parameter group.
-
-        Parameters
-        ----------
-        model :
-            A glotaran model.
-        parameter : ParameterGroup
-            The parameter group to fill from.
-        """
-        item = copy.deepcopy(self)
-        for name in self._glotaran_properties:
-            prop = getattr(self.__class__, name)
-            value = getattr(self, name)
-            value = prop.glotaran_fill(value, model, parameters)
-            setattr(item, name, value)
-        return item
-
-    return fill
-
-
-def _get_state_factory(cls):
-    @wrap_func_as_method(cls)
-    def get_state(self) -> cls:
-        return tuple(getattr(self, name) for name in self._glotaran_properties)
-
-    return get_state
-
-
-def _set_state_factory(cls):
-    @wrap_func_as_method(cls)
-    def set_state(self, state) -> cls:
-        for i, name in enumerate(self._glotaran_properties):
-            setattr(self, name, state[i])
-
-    return set_state
-
-
-def _markdown_factory(cls):
-    @wrap_func_as_method(cls, name="markdown")
-    def mprint_item(
-        self, all_parameters: ParameterGroup = None, initial_parameters: ParameterGroup = None
-    ) -> MarkdownStr:
-        f"""Returns a string with the {cls.__name__} formatted in markdown."""
-
-        md = "\n"
-        if self._glotaran_has_label:
-            md = f"**{self.label}**"
-            if hasattr(self, "type"):
-                md += f" ({self.type})"
-            md += ":\n"
-
-        elif hasattr(self, "type"):
-            md = f"**{self.type}**:\n"
-
-        for name in self._glotaran_properties:
-            prop = getattr(self.__class__, name)
-            value = getattr(self, name)
-            if value is None:
-                continue
-            property_md = indent(
-                f"* *{name.replace('_', ' ').title()}*: "
-                f"{prop.glotaran_value_as_markdown(value,all_parameters, initial_parameters)}\n",
-                "  ",
-            )
-
-            md += property_md
-
-        return MarkdownStr(md)
-
-    return mprint_item
-
-
-def _get_parameter_labels_factory(cls):
-    @wrap_func_as_method(cls, name="get_parameter_labels")
-    def get_parameter_labels(self) -> list[str]:
-
-        parameter_labels = []
-
-        for name in self._glotaran_properties:
-            prop = getattr(self.__class__, name)
-            value = getattr(self, name)
-            parameter_labels += prop.glotaran_get_parameter_labels(value)
-
-        return parameter_labels
-
-    return get_parameter_labels
+    alias: str | None = None,
+    default: any = NOTHING,
+    validator: Callable[[ModelItem, Model, ParameterGroup | None], list[ItemIssue]] | None = None,
+) -> Attribute:
+    metadata = {}
+    if alias is not None:
+        metadata[META_ALIAS] = alias
+    if validator is not None:
+        metadata[META_VALIDATOR] = validator
+    return field(default=default, metadata=metadata)
