+++ glotaran/model/test/test_item.py
@@ -1,17 +1,17 @@
 from attrs import fields
 
-from glotaran.model_new.item import ModelItem
-from glotaran.model_new.item import ModelItemType
-from glotaran.model_new.item import ParameterType
-from glotaran.model_new.item import fill_item
-from glotaran.model_new.item import get_item_model_issues
-from glotaran.model_new.item import get_item_parameter_issues
-from glotaran.model_new.item import item
-from glotaran.model_new.item import model_attributes
-from glotaran.model_new.item import strip_type_and_structure_from_attribute
-from glotaran.model_new.megacomplex import Megacomplex
-from glotaran.model_new.megacomplex import megacomplex
-from glotaran.model_new.model import Model
+from glotaran.model.item import ModelItem
+from glotaran.model.item import ModelItemType
+from glotaran.model.item import ParameterType
+from glotaran.model.item import fill_item
+from glotaran.model.item import get_item_model_issues
+from glotaran.model.item import get_item_parameter_issues
+from glotaran.model.item import item
+from glotaran.model.item import model_attributes
+from glotaran.model.item import strip_type_and_structure_from_attribute
+from glotaran.model.megacomplex import Megacomplex
+from glotaran.model.megacomplex import megacomplex
+from glotaran.model.model import Model
 from glotaran.parameter import Parameter
 from glotaran.parameter import ParameterGroup
