--- glotaran/model_new/test/test_dataset_model.py
@@ -1,65 +1,0 @@
-"""Tests for glotaran.model.dataset_model.DatasetModel"""
-from __future__ import annotations
-
-from glotaran.model_new.item import get_item_issues
-from glotaran.model_new.megacomplex import Megacomplex
-from glotaran.model_new.megacomplex import megacomplex
-from glotaran.model_new.model import Model
-
-
-@megacomplex()
-class MockMegacomplexNonUniqueExclusive(Megacomplex):
-    type: str = "test_megacomplex_not_exclusive_unique"
-
-
-@megacomplex(exclusive=True)
-class MockMegacomplexExclusive(Megacomplex):
-    type: str = "test_megacomplex_exclusive"
-
-
-@megacomplex(unique=True)
-class MockMegacomplexUnique(Megacomplex):
-    type: str = "test_megacomplex_unique"
-
-
-def test_get_issues_datasetmodel():
-    mcls = Model.create_class_from_megacomplexes(
-        [
-            MockMegacomplexNonUniqueExclusive,
-            MockMegacomplexExclusive,
-            MockMegacomplexUnique,
-        ]
-    )
-    m = mcls(
-        megacomplex={
-            "m": {"type": "test_megacomplex_not_exclusive_unique"},
-            "m_exclusive": {"type": "test_megacomplex_exclusive"},
-            "m_unique": {"type": "test_megacomplex_unique"},
-        },
-        dataset={
-            "ok": {"megacomplex": ["m"]},
-            "exclusive": {"megacomplex": ["m_exclusive", "m_exclusive"]},
-            "unique": {"megacomplex": ["m", "m_unique"]},
-        },
-    )
-
-    assert len(get_item_issues(item=m.dataset["ok"], model=m)) == 0
-    assert len(get_item_issues(item=m.dataset["exclusive"], model=m)) == 2
-    assert len(get_item_issues(item=m.dataset["unique"], model=m)) == 1
-
-    m = mcls(
-        megacomplex={
-            "m": {"type": "test_megacomplex_not_exclusive_unique"},
-            "m_exclusive": {"type": "test_megacomplex_exclusive"},
-            "m_unique": {"type": "test_megacomplex_unique"},
-        },
-        dataset={
-            "ok": {"megacomplex": [], "global_megacomplex": ["m"]},
-            "exclusive": {"megacomplex": [], "global_megacomplex": ["m_exclusive", "m_exclusive"]},
-            "unique": {"megacomplex": [], "global_megacomplex": ["m", "m_unique"]},
-        },
-    )
-
-    assert len(get_item_issues(item=m.dataset["ok"], model=m)) == 0
-    assert len(get_item_issues(item=m.dataset["exclusive"], model=m)) == 2
-    assert len(get_item_issues(item=m.dataset["unique"], model=m)) == 1
