+++ setup.cfg
@@ -61,13 +61,6 @@ glotaran.plugins.data_io =
     ascii = glotaran.builtin.io.ascii.wavelength_time_explicit_file
     sdt = glotaran.builtin.io.sdt.sdt_file_reader
     nc = glotaran.builtin.io.netCDF.netCDF
-glotaran.plugins.megacomplexes =
-    baseline = glotaran.builtin.megacomplexes.baseline
-    clp_guide = glotaran.builtin.megacomplexes.clp_guide
-    coherent_artifact = glotaran.builtin.megacomplexes.coherent_artifact
-    damped_oscillation = glotaran.builtin.megacomplexes.damped_oscillation
-    decay = glotaran.builtin.megacomplexes.decay
-    spectral = glotaran.builtin.megacomplexes.spectral
 glotaran.plugins.project_io =
     yml = glotaran.builtin.io.yml.yml
     csv = glotaran.builtin.io.pandas.csv
